introduce plugin system
still bugs in battery_charge_current
